refactor(container): mv UserBanner b-> UserDigest
