--- src/containers/banner/UserBanner/tests/store.test.js
@@ -1,10 +1,0 @@
-/*
- * UserBanner store test
- *
- */
-
-// import UserBanner from '../index'
-
-it('TODO: store test UserBanner', () => {
-  expect(1 + 1).toBe(2)
-})
