+++ src/containers/digest/UserDigest/tests/index.test.js
@@ -1,9 +1,9 @@
 // import React from 'react'
 // import { shallow } from 'enzyme'
 
-// import UserBanner from '../index'
+// import UserDigest from '../index'
 
-describe('TODO <UserBanner />', () => {
+describe('TODO <UserDigest />', () => {
   it('Expect to have unit tests specified', () => {
     expect(true).toEqual(true)
   })
