+++ src/containers/digest/UserDigest/tests/store.test.js
@@ -1,0 +1,10 @@
+/*
+ * UserDigest store test
+ *
+ */
+
+// import UserDigest from '../index'
+
+it('TODO: store test UserDigest', () => {
+  expect(1 + 1).toBe(2)
+})
